backend/requirements.txt
Flask
Pillow
firebase-admin
google-genai
google-cloud-storage
google-cloud-firestore
google-cloud-texttospeech
google-cloud-secret-manager==2.16.0
gunicorn
dotenv
flask-CORS
dotenv